requirements.txt
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
requests>=2.32.3
pydantic==2.6.1
python-multipart==0.0.6
deepseek-ai==0.0.1
SQLAlchemy==2.0.25
aiohttp==3.9.3
python-jose==3.3.0
passlib==1.7.4
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
aiosqlite==0.19.0 